Refactor config.py to optionally read DB and Flask host config from environment variables.
If value isn't set in the environment, defaults are provided with an exception in ProdConf, where exception should be thrown of no DB settings are provided by environment var.
ProdConfig and DevConfig now inherit from BaseConfig.  This required refactoring "config" dictionary in places_api.py to be a class rather than dictionary, as vars() doesn't seem to handle inheritance.

Resolves #28